refactor: unify VDropdown — replace old Menu/Picker with VMenuPanel, remove VFilterDropdown
- Rewrite VDropdown to use VMenuPanel on macOS (VMenu with proper
  z-ordering, click-outside dismiss, toggle-to-close)
- Delete VFilterDropdown — merged into the unified VDropdown
- Add excludeRect to VMenuPanel so trigger clicks toggle instead of
  reopening the menu
- Remove error state (gallery-only, never used in production)
- Remove size variants (always regular 32pt)
- Keep backwards-compatible tuple init for existing call sites
- Update gallery to remove error/size sections
