+++ clients/macos/vellum-assistant/Features/Settings/SettingsDeveloperTab.swift
@@ -573,7 +573,6 @@ struct SettingsDeveloperTab: View {
                         placeholder: "",
                         selection: $selectedAssistantId,
                         options: awakeAssistants.map { (label: displayLabel(for: $0), value: $0.assistantId) },
-                        size: .small,
                         maxWidth: 200
                     )
                 }
